refactor: remove redundant code and integrate minyoung's improvde code with previous code
