+++ .gitignore
@@ -1,0 +1,4 @@
+temp
+data/test_db
+data/test_data
+.env
