--- FinQA/0602_mny/main.py
@@ -1,6 +1,0 @@
-def main():
-    print("Hello from mcp!")
-
-
-if __name__ == "__main__":
-    main()
