--- FinQA/0602_mny/pyproject.toml
@@ -1,10 +1,0 @@
-[project]
-name = "mcp_test"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.13"
-dependencies = [
-    "httpx>=0.28.1",
-    "mcp[cli]>=1.6.0",
-]
