--- FinQA/0602_mny/requirements.txt
@@ -1,152 +1,0 @@
-aiohappyeyeballs==2.6.1
-aiohttp==3.11.18
-aiosignal==1.3.2
-altair==5.5.0
-annotated-types==0.7.0
-anyio==4.9.0
-asgiref==3.8.1
-attrs==25.3.0
-backoff==2.2.1
-bcrypt==4.3.0
-blinker==1.9.0
-build==1.2.2.post1
-cachetools==5.5.2
-certifi==2025.1.31
-charset-normalizer==3.4.2
-chroma-hnswlib==0.7.6
-chromadb==0.6.3
-click==8.1.8
-coloredlogs==15.0.1
-dataclasses-json==0.6.7
-deprecated==1.2.18
-distro==1.9.0
-durationpy==0.9
-fastapi==0.115.12
-filelock==3.18.0
-flatbuffers==25.2.10
-frozenlist==1.6.0
-fsspec==2025.3.2
-gitdb==4.0.12
-gitpython==3.1.44
-google-auth==2.40.1
-googleapis-common-protos==1.70.0
-greenlet==3.2.1
-grpcio==1.71.0
-h11==0.14.0
-httpcore==1.0.8
-httptools==0.6.4
-httpx==0.28.1
-httpx-sse==0.4.0
-huggingface-hub==0.30.2
-humanfriendly==10.0
-idna==3.10
-importlib-metadata==8.6.1
-importlib-resources==6.5.2
-jinja2==3.1.6
-jiter==0.9.0
-jsonpatch==1.33
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2025.4.1
-kubernetes==32.0.1
-langchain==0.3.25
-langchain-chroma==0.2.3
-langchain-community==0.3.23
-langchain-core==0.3.58
-langchain-mcp-adapters==0.0.10
-langchain-openai==0.3.16
-langchain-text-splitters==0.3.8
-langgraph==0.4.2
-langgraph-checkpoint==2.0.25
-langgraph-prebuilt==0.1.8
-langgraph-sdk==0.1.66
-langsmith==0.3.42
-markdown-it-py==3.0.0
-markupsafe==3.0.2
-marshmallow==3.26.1
-mcp==1.7.1
-mdurl==0.1.2
-mmh3==5.1.0
-mpmath==1.3.0
-multidict==6.4.3
-mypy-extensions==1.1.0
-narwhals==1.38.2
-nest-asyncio==1.6.0
-numpy==2.2.5
-oauthlib==3.2.2
-onnxruntime==1.21.1
-openai==1.77.0
-opentelemetry-api==1.32.1
-opentelemetry-exporter-otlp-proto-common==1.32.1
-opentelemetry-exporter-otlp-proto-grpc==1.32.1
-opentelemetry-instrumentation==0.53b1
-opentelemetry-instrumentation-asgi==0.53b1
-opentelemetry-instrumentation-fastapi==0.53b1
-opentelemetry-proto==1.32.1
-opentelemetry-sdk==1.32.1
-opentelemetry-semantic-conventions==0.53b1
-opentelemetry-util-http==0.53b1
-orjson==3.10.18
-ormsgpack==1.9.1
-overrides==7.7.0
-packaging==24.2
-pandas==2.2.3
-pillow==11.2.1
-posthog==4.0.1
-propcache==0.3.1
-protobuf==5.29.4
-pyarrow==20.0.0
-pyasn1==0.6.1
-pyasn1-modules==0.4.2
-pydantic==2.11.3
-pydantic-core==2.33.1
-pydantic-settings==2.8.1
-pydeck==0.9.1
-pygments==2.19.1
-pypika==0.48.9
-pyproject-hooks==1.2.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
-python-multipart==0.0.20
-pytz==2025.2
-pyyaml==6.0.2
-referencing==0.36.2
-regex==2024.11.6
-requests==2.32.3
-requests-oauthlib==2.0.0
-requests-toolbelt==1.0.0
-rich==14.0.0
-rpds-py==0.24.0
-rsa==4.9.1
-shellingham==1.5.4
-six==1.17.0
-smmap==5.0.2
-sniffio==1.3.1
-sqlalchemy==2.0.40
-sse-starlette==2.2.1
-starlette==0.46.2
-streamlit==1.45.0
-sympy==1.14.0
-tenacity==9.1.2
-tiktoken==0.9.0
-tokenizers==0.21.1
-toml==0.10.2
-tornado==6.4.2
-tqdm==4.67.1
-typer==0.15.2
-typing-extensions==4.13.2
-typing-inspect==0.9.0
-typing-inspection==0.4.0
-tzdata==2025.2
-urllib3==2.4.0
-uvicorn==0.34.1
-uvloop==0.21.0
-watchdog==6.0.0
-watchfiles==1.0.5
-websocket-client==1.8.0
-websockets==15.0.1
-wrapt==1.17.2
-xxhash==3.5.0
-yarl==1.20.0
-zipp==3.21.0
-zstandard==0.23.0
