+++ FinQA/results/v0530/results_thoughts_fy.json
