+++ FinQA/results/v0530/results_thoughts_v2.json
