+++ FinQA/results/v0530/results_thoughts_v9.json
