+++ FinQA/score.py
@@ -4,6 +4,7 @@ import json
 from pydantic import BaseModel, Field
 from dotenv import load_dotenv, find_dotenv
 import os
+from tqdm import tqdm
 
 _ = load_dotenv(find_dotenv())
 
@@ -11,11 +12,12 @@ OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
 
 llm = ChatOpenAI(model="gpt-4o-mini", temperature=0.0, api_key=OPENAI_API_KEY)
 
-with open('./data/qa_dict.json', 'r') as f:
+with open('./data/qa_dict_levels.json', 'r') as f:
     qa_dict = json.load(f)
 
-with open('./data/results.json', 'r') as f:
+with open('./results/v0602/results_thoughts_v1.json', 'r') as f:
     results = json.load(f)
+    
 
 score_answer_prompt = PromptTemplate(
     input_variables=["question", "answer", "response"],
@@ -42,16 +44,27 @@ class Score(BaseModel):
 score_answer_chain = score_answer_prompt | llm.with_structured_output(Score)
 
 correct = 0
-for i, item in enumerate(qa_dict):
-  response = results[i]['Output']
+for i, item in tqdm(enumerate(qa_dict), desc="Evaluating answers"):
+  response = results[i]['final_answer']
   score = score_answer_chain.invoke({'question': item['Question'], 'answer': item['Answer'], 'response': response}).score
   correct += score
+  del(results[i]['question'])
+  del(results[i]['trace'])
+  results[i]['Question'] = item['Question']
+  results[i]['System Answer'] = results[i]['final_answer']
+  del(results[i]['final_answer'])
+  results[i]['True Answer'] = item['Answer']
+  results[i]['Level'] = item['Level']
   results[i]['Score'] = score
 
 accuracy = correct / len(qa_dict)
 
 print(f"Accuracy: {accuracy}")
 
-with open('./data/results_with_score.json', 'w') as f:
+# Create the output directory if it doesn't exist
+output_dir = "./scores/v0602"
+os.makedirs(output_dir, exist_ok=True)
+
+with open('./scores/v0602/results_with_score_v1.json', 'w') as f:
     json.dump(results, f, indent=4)
-    print(f"Results with score saved to results_with_score.json")
+    print(f"Results with score saved")
